CLI: moved utility functions to /utils (read below):
- moving the 2 utils getMigrationsExtensions, getStubPath had dependencies
- the depended utils had to be moved over as well and their libs imported
- utils moved: exit, success, resolveEnvironmentConfig, checkLocalModule
